backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
asyncpg>=0.29.0
redis[asyncio]>=5.0.0
pydantic-settings>=2.2.0
openai>=1.30.0
anthropic>=0.26.0
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-instrumentation-fastapi>=0.45b0
opentelemetry-exporter-otlp>=1.24.0
prometheus-client>=0.20.0
python-dotenv>=1.0.0
sse-starlette>=2.1.0
httpx>=0.27.0
pypdfium2>=4.27.0
pdfplumber>=0.11.0
pytesseract>=0.3.10
python-docx>=1.1.0
pandas>=2.2.0
tabulate>=0.9.0
trafilatura>=1.9.0
tiktoken>=0.7.0
Pillow>=10.3.0
arq>=0.26.0
mlflow>=2.13.0
scipy>=1.13.0
sentence-transformers>=3.0.0
